Added option for providing user specified stoplist for content extraction in html files.
This SHOULD be provided by the user. The stoplist in the stock jusText is to small for reliably extracting security related data in text fields.

The current stoplist used in testing contains all the "Words" from scio configs (Threat actors, tools etc) as well as a list of ~370.000 english words. (The jusText English stoplist is only ~450 words.)