Refactor of feeds return values and API bugfix
Refactored the feeds dataflow to retuyrn a list of metadata in a map instead of
a list of filenames. 'filename' is now a key in this metadata map.

The feeds upload now honors the API back-pressure

Added 'uri' to the API document type

BUGFIX: max_current_jobs_ready did not check for tube existence before stat'ing
the tube leading to a 500 error after beanstalk restart when no queues was
created.